backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy>=1.4
asyncpg
greenlet
python-dotenv
joblib
pandas
numpy
scikit-learn
torch
transformers
accelerate>=0.26.0
datasets
lightgbm
statsmodels
shap
